pyproject.toml
[project]
name = "hack-interview"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.17"
dependencies = [
    "annotated-types==0.6.0 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "anyio==3.7.1 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "certifi==2023.11.17 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "cffi==1.16.0 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "colorama==0.4.6 ; python_full_version >= '3.10' and python_full_version < '4.0' and sys_platform == 'win32'",
    "distro==1.8.0 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "exceptiongroup==1.1.3 ; python_full_version == '3.10.*'",
    "freesimplegui>=5.2.0.post1",
    "h11==0.14.0 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "httpcore==1.0.2 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "httpx==0.25.1 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "idna==3.4 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "loguru==0.7.2 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "numpy==1.26.2 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "openai==1.3.3 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "pillow>=11.2.1",
    "pycparser==2.21 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "pydantic==2.5.1 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "pydantic-core==2.14.3 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "python-dotenv==1.0.0 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "sniffio==1.3.0 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "sounddevice==0.4.6 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "soundfile==0.12.1 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "tqdm==4.66.1 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "typing-extensions==4.8.0 ; python_full_version >= '3.10' and python_full_version < '4.0'",
    "win32-setctime==1.1.0 ; python_full_version >= '3.10' and python_full_version < '4.0' and sys_platform == 'win32'",
]
